Makes attention and folio reusable
