+++ ppr-ui/src/components/common/index.ts
@@ -28,3 +28,5 @@ export { default as ContactInformation } from './ContactInformation.vue'
 export { default as ContactUsToggle } from './ContactUsToggle.vue'
 export { default as SubProductSelector } from './SubProductSelector.vue'
 export { default as QsAccessBtn } from './QsAccessBtn.vue'
+export { default as Attention } from './Attention.vue'
+export { default as FolioOrReferenceNumber } from './FolioOrReferenceNumber.vue'
