+++ ppr-ui/src/components/mhrRegistration/ReviewConfirm/index.ts
@@ -2,5 +2,3 @@ export { default as YourHomeReview } from './YourHomeReview.vue'
 export { default as SubmittingPartyReview } from './SubmittingPartyReview.vue'
 export { default as HomeLocationReview } from './HomeLocationReview.vue'
 export { default as HomeOwnersReview } from './HomeOwnersReview.vue'
-export { default as Attention } from './Attention.vue'
-export { default as FolioOrReferenceNumber } from './FolioOrReferenceNumber.vue'
